Fixing sum square calculation
It should fit in a double, not in a potentially unsigned int

+++ cryptominisat3/avgcalc.h
@@ -57,7 +57,7 @@ public:
 
     void push(const T x) {
         sum += x;
-        sumSqare += x*x;
+        sumSqare += (double)x*(double)x;
         num++;
 
         max = std::max(max, x);
